🚚 Consolidate type defintions
