+++ src/types/index.ts
@@ -34,3 +34,13 @@ export type {
 export type {
   ZodAlertType,
 } from '@/types/zodShared';
+
+export type ValueOf<T> = T[keyof T];
+
+export type Nullable<T> = null | T | undefined;
+
+export type NonNullable<T> = T extends null | undefined ? never : T;
+
+export type DictKey = number | string | symbol;
+
+export type Dict<K extends DictKey = string, T = any> = Record<K, T>;
