+++ src/types/zodFlaw.ts
@@ -2,7 +2,7 @@ import { z } from 'zod';
 import { DateTime } from 'luxon';
 
 import { cveRegex } from '@/utils/helpers';
-import type { ValueOf } from '@/utils/typeHelpers';
+import type { ValueOf } from '@/types';
 
 import {
   MajorIncidentStateEnum,
