--- src/utils/typeHelpers.ts
@@ -1,5 +1,0 @@
-export type ValueOf<T> = T[keyof T];
-
-export type Nullable<T> = null | T | undefined;
-
-export type NonNullable<T> = T extends null | undefined ? never : T;
